requirements.txt
streamlit
pytesseract
pillow
pymupdf
python-docx
pymongo
pandas
openpyxl
